requirements.txt
Fabric==1.10.0
PyYAML==3.11
pycrypto==2.6.1
Flask==0.11.1
flask-restful==0.3.5
gevent==1.1.2
mock==1.3.0
nose==1.3.7